requirements.txt
flask
pyVoIP
audioop-lts
gunicorn
requests
PySocks
